+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v1.3.0
 * Added parsing for characters who have completed Moogle Beast Tribe to Rank 7
 * Added parsing for characters who have completed Vanu Vanu Beast Tribe to Rank 7
 * Added parsing for characters who have completed Vath Beast Tribe to Rank 7
+* [Bugfix] 180 days and 90 day subscriptions were returing wrong value
 
 v1.2.1
 ======